add order's policies

+++ app/Http/Controllers/Admin/Market/OrderController.php
@@ -10,6 +10,7 @@ class OrderController extends Controller
 {
     public function index()
     {
+        $this->authorize('index', Order::class);
         $page = "همه سفارشات";
         $orders = Order::all();
         return view('admin.market.order.index', compact('orders', 'page'));
@@ -17,6 +18,7 @@ class OrderController extends Controller
 
     public function newOrders()
     {
+        $this->authorize('index', Order::class);
         $page = "سفارشات جدید";
         $orders = Order::where('order_status', 0)->get();
         foreach ($orders as $order) {
@@ -27,6 +29,7 @@ class OrderController extends Controller
 
     public function processing()
     {
+        $this->authorize('index', Order::class);
         $page = "سفارشات در حال پردازش";
         $orders = Order::where('order_status', 1)->get();
         return view('admin.market.order.index', compact('orders', 'page'));
@@ -34,6 +37,7 @@ class OrderController extends Controller
 
     public function delivering()
     {
+        $this->authorize('index', Order::class);
         $page = "سفارشات در حال ارسال";
         $orders = Order::where('order_status', 2)->get();
         return view('admin.market.order.index', compact('orders', 'page'));
@@ -41,6 +45,7 @@ class OrderController extends Controller
 
     public function unpaid()
     {
+        $this->authorize('index', Order::class);
         $page = "سفارشات پرداخت نشده";
         $orders = Order::where('payment_status', 0)->get();
         return view('admin.market.order.index', compact('orders', 'page'));
@@ -48,6 +53,7 @@ class OrderController extends Controller
 
     public function expired()
     {
+        $this->authorize('index', Order::class);
         $page = "سفارشات لغو شده";
         $orders = Order::where('order_status', 4)->get();
         return view('admin.market.order.index', compact('orders', 'page'));
@@ -55,6 +61,7 @@ class OrderController extends Controller
 
     public function returned()
     {
+        $this->authorize('index', Order::class);
         $page = "سفارشات مرجوعی";
         $orders = Order::where('order_status', 5)->get();
         return view('admin.market.order.index', compact('orders', 'page'));
@@ -62,6 +69,7 @@ class OrderController extends Controller
 
     public function changeStatus(Order $order)
     {
+        $this->authorize('update', Order::class);
         $order->order_status = ($order->order_status + 1) > 5 ? 0 : $order->order_status + 1;
         $order->save();
 
@@ -69,6 +77,7 @@ class OrderController extends Controller
     }
     public function changeDeliveryStatus(Order $order)
     {
+        $this->authorize('update', Order::class);
         $order->delivery_status = ($order->delivery_status + 1) > 2 ? 0 : $order->delivery_status + 1;
         $order->save();
 
@@ -77,6 +86,7 @@ class OrderController extends Controller
 
     public function destroy(Order $order)
     {
+        $this->authorize('delete', Order::class);
         $order->delete();
 
         return redirect()->back()->with('alertify-success', 'سفارش موردنظر با موفقیت حذف شد');
@@ -84,6 +94,7 @@ class OrderController extends Controller
 
     public function show(Order $order)
     {
+        $this->authorize('index', Order::class);
         return view('admin.market.order.show', compact('order'));
     }
 }
